pyproject.toml
[project]
name = "greek-room-api"
version = "0.1.0"
description = "Greek Room API and MCP endpoints with Streamlit Chat UI powered by LLM"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.118.0",
    "fastmcp>=2.12.4",
    "loguru>=0.7.3",
    "openai>=2.2.0",
    "predictionguard>=2.9.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "regex>=2025.7.34",
    "streamlit>=1.50.0",
    "unicodeblock>=0.3.1",
    "uvicorn>=0.37.0",
    "wildebeest-nlp==0.9.2",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
]
